# Fix memory consumption when multiple models are used