--- a/solution_move_stops_generator.go
+++ b/solution_move_stops_generator.go
@@ -2,10 +2,6 @@
 
 package nextroute
 
-import (
-	"github.com/nextmv-io/nextroute/common"
-)
-
 // SolutionMoveStopsGeneratorChannel generates all possible moves for a given
 // vehicle and plan unit.
 //
@@ -101,12 +97,8 @@ func SolutionMoveStopsGenerator(
 	preAllocatedMoveContainer *PreAllocatedMoveContainer,
 	shouldStop func() bool,
 ) {
-	source := common.Map(stops, func(stop SolutionStop) SolutionStop {
-		return stop
-	})
-	target := common.Map(vehicle.SolutionStops(), func(stop SolutionStop) SolutionStop {
-		return stop
-	})
+	source := stops
+	target := vehicle.SolutionStops()
 	m := preAllocatedMoveContainer.singleStopPosSolutionMoveStop
 	m.(*solutionMoveStopsImpl).reset()
 	m.(*solutionMoveStopsImpl).planUnit = planUnit