--- a/solution_test.go
+++ b/solution_test.go
@@ -39,3 +39,38 @@ func BenchmarkAllocationsSolution(b *testing.B) {
 		}
 	}
 }
+
+// TestLimitAllocations tests the number of allocations in the solution creation.
+// We want to ensure that the number of allocations is limited and does not grow
+// accidentally.
+func TestLimitAllocations(t *testing.T) {
+	model, err := createModel(singleVehiclePlanSequenceModel())
+	if err != nil {
+		t.Error(err)
+	}
+
+	maximum := nextroute.NewVehicleTypeDurationExpression(
+		"maximum duration",
+		3*time.Minute,
+	)
+	expression := nextroute.NewStopExpression("test", 2.0)
+
+	cnstr, err := nextroute.NewMaximum(expression, maximum)
+	if err != nil {
+		t.Error(err)
+	}
+
+	err = model.AddConstraint(cnstr)
+	if err != nil {
+		t.Error(err)
+	}
+	allocs := testing.AllocsPerRun(2, func() {
+		_, err = nextroute.NewSolution(model)
+		if err != nil {
+			t.Fatal(err)
+		}
+	})
+	if allocs > 66 {
+		t.Errorf("expected 66 allocations, got %v", allocs)
+	}
+}