This fixes a memory leak (in a way) when using multiple models within the same program with expressions.

The problem is that every expression gets an ID which is unique within the package. It uses a package global counter.
When creating a new solution we create 2 two dimensional slices with one dimension being the maximum expression ID. Every time we create anywhere a new model, this maximum expression ID increases. 

For example:
Creating a model in a loop with `n` expressions and `m` stops and solving it will slowly consume all memory since with each iteration the memory consumed when creating the first solution grows by `2 * n * m bytes`.

The benchmark test show cases that. Here I do exactly that with a very small model and even that shows already the effect. This change though comes at a cost of using a map data structure internally for now, which is more expensive.

```
name                   old time/op    new time/op    delta
AllocationsSolution-8     481µs ±30%      44µs ± 1%  -90.95%  (p=0.000 n=9+9)

name                   old alloc/op   new alloc/op   delta
AllocationsSolution-8    6.45MB ±77%    0.03MB ± 0%  -99.58%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
AllocationsSolution-8       189 ± 1%       188 ± 0%   -0.63%  (p=0.000 n=10+10)
```