@dirkschumacher 2024-05-19
@davidrijsman not super straightforward, but I added enough to compensate for the slower access times.

If we merge this then the bug is fixed and we have more time to find a better solution that uses slices.